// packages/app/app/components/map-view-directions/types.ts
export interface MapViewDirectionsState {
  coordinates: { latitude: number, longitude: number }[] | null
  distance: number | null
  duration: number | null
}